+++ tests/integration/test_recent_performance_samples.py
@@ -4,7 +4,7 @@ import time
 
 from pytest import fixture, mark
 
-from .utils import assert_valid_response
+from ..utils import assert_valid_response
 
 
 @fixture(scope="session")